flow: use a specialized linearizer
This patch enables the linearization infrastructure of the research
codes in eWoms to diverge from the production ones while guaranteeing
stability for the latter. The actual new linearizer class for flow
currently corresponds to the one of eWoms master, but with slightly
more extensive usage of openMP.

Another purpose of the patch is to provide a blueprint of how
diverging requirements of research- and production codes can be
handled in the future: In this approach, the affected classes are
simply forked off and replace their original by overwriting the
respective eWoms property; in the particular case of `flow`, this
approach allows to add ECL-pecific code in generic classes which
cannot be easily changed without severely disturbing the eWoms-based
research simulators.