make BlackoilWellModel::numComponents return consistent values.
The return value is used to resize a vector that is used unconditionally
(i.e. whether there are wells or not) elsewhere, e.g. in
computeAverageFormationFactor, called from computeWellPotentials
called from timeStepSucceeded. Hence it has to return the same value
whether on all processes whether there wells (globally or locally) or
not.

This is done by this commit that introduces global_num_phases_ that
is the max of all phases on any process.