Skip well computations if there are no wells.
These computations would need correctly sized containers (B_avg).
If there are no we cannot determine this correct size from the wells,
therefore we choose to skip those computations as they are not needed
anyway and just cost time.

The alternative would been to count the number of active phases.