add custom target 'opm-simulators_prepare'
this acts as a syncronization point for the super-build.
the ebos objects do not depend directly on opm-common, and thus
they are can be built before the generated headers, which are
required, are in place. we can now use this to ensure the
generated headers are in place before proceeding.