Refactor flowEbosBlackoilMain()
Currently, the Python run() calls flowEbosBlackoilMain() to start the
simulation (which calls execute() in FlowMainEbos). The Python
step_init() method should also do the setup in flowEbosBlackoilMain()
but it should not call execute() directly. In order to avoid code
duplication for run() and step_init(), the initialization part of
flowEbosBlackoilMain() is isolated into a flowEbosBlackoilMain_init()