Python bindings for the blackoil simulator.
NOTE: This PR depends on PR #2516 which should be merged first.

A rewrite of the Python bindings for the blackoil simulator using
pybind11 as introduced in PR #2127. The new version uses the refactored
flow.cpp introduced in PR #2516 and thus avoids code duplication in
simulators.cpp.

This PR will be the starting point for implementing the Python bindings
described in PR #2403.