Refactor flow_ebos_blackoil.cpp.
NOTE: this pull request depends on PR #2518 which should be merged
first.

In this PR, we return to the list of PRs presented in Pull #2403,
starting with the first two:

- PR #2438 : Refactor run() into prepareRun_(), and
- PR #2439 : refactor flowEbosBlackoilMain().

The first, #2438, has already been handled by the refactoring of
flow.cpp in PR #2516. The second, #2439, has diverged to much from the
current master, so I decided to create a new PR (this one) instead of
trying to force-adjust it.

We would like to add a step() method to simulators.cpp, that will let
a Python script advance the simulator one report step at a time. Before
calling step(), the Python script will have to call a step_init() method
that initializes the simulator, similar to what is done in run() now.

Currently, the Python run() calls runDynamic() in Main.hpp, which calls
dispatchDynamic_() which again calls flowEbosBlackoilMain() to start the
simulation (which calls execute() in class FlowMainEbos). The Python
step_init() method should also do the setup in flowEbosBlackoilMain()
but it should not call execute() directly. In order to avoid code
duplication for run() and step_init(), the initialization part of
flowEbosBlackoilMain() is isolated into a flowEbosBlackoilMainInit()
method.

The step_init() method (to be implemented in a later PR) will now call
initFlowEbosBlackoil() in Main.hpp to initialize the simulator.