Refactor runSimulator().
A resubmission of commit b25f489 in PR #2403 and PR #2441 to work with
the current master.

Continues the work in PR #2619 to refactor FlowMainEbos.hpp to work with
the Python bindings.

We need to refactor runSimulator() to avoid code duplication when
executeStepInit() is implemented (see later commit). Here, runSimulator()
is refactored into a runSimulatorInitOrRun() that takes a callback
function. When runSimulatorInit() is implemented it will pass a different
callback that only initializes the simulator. Currently, runSimulator()
passes the callback runSimulatorRunCallback_() which runs the whole
simulation. The code outputting the simulation summary in runSimulator()
is also refactored into a runSimulatorAfterSim_() method.