Added a class with information and comunicator for parallel wells.
BlackoilWellModel now stores an instance of this class for each
well. Inside that class there is a custom communicator that only
contains ranks that will have local cells perforated by the well.
This will be used in the application of the distributed.

This is another small step in the direction of distrbuted wells,
but it should be safe to merge this (note creation of the custom
communicators is a collective operation in MPI but done only once).