Simplify computation of connection pressure delta.
Previously we did compute the delta between two perforations
and then used a partial sum. That seemed to be a rather complex
algorithm to calculate the delta between a connection and the
reference depth.
This commit now simply computes the latter without an intermediary
step (difference between two perforations) and hopefully is more
clear.

It is also more suitable for distributed wells.